중복 유틸리티 함수 - 공통 모듈로 추출
