+++ src/utils/dateUtils.js
@@ -34,7 +34,7 @@ export const formatDateToMonthDay = (dateString) => {
 // 시간 형식 변환 함수 ({hour, minute} 또는 "HH:mm:ss" -> "HH:MM")
 export const formatTime = (timeObj) => {
   if (!timeObj) return "";
-  
+
   // 문자열 형식인 경우 ("HH:mm:ss" -> "HH:mm")
   if (typeof timeObj === "string") {
     try {
@@ -46,27 +46,27 @@ export const formatTime = (timeObj) => {
       return "";
     }
   }
-  
+
   // 객체 형식인 경우 ({hour, minute} -> "HH:mm")
   if (timeObj.hour !== undefined && timeObj.minute !== undefined) {
     const hour = String(timeObj.hour).padStart(2, "0");
     const minute = String(timeObj.minute).padStart(2, "0");
     return `${hour}:${minute}`;
   }
-  
+
   return "";
 };
 
 // 날짜 문자열을 숫자와 요일로 변환 ("2025-12-17" -> { date: 17, day: "수" })
 export const parseWorkDate = (dateString) => {
   if (!dateString) return { date: 0, day: "" };
-  
+
   try {
     const [year, month, day] = dateString.split("-").map(Number);
     const date = new Date(year, month - 1, day);
     const dayIndex = date.getDay();
     const dayLabels = ["일", "월", "화", "수", "목", "금", "토"];
-    
+
     return {
       date: day,
       day: dayLabels[dayIndex] || "",
@@ -76,3 +76,21 @@ export const parseWorkDate = (dateString) => {
   }
 };
 
+
+/**
+ * 서버 시간 포맷을 한국어 형식으로 변환
+ * @param {string} dateString - ISO 형식의 날짜 문자열
+ * @returns {string} 한국어 형식의 날짜 문자열
+ */
+export const formatDateTime = (dateString) => {
+  if (!dateString) return "";
+
+  const date = new Date(dateString);
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  const hours = String(date.getHours()).padStart(2, "0");
+  const minutes = String(date.getMinutes()).padStart(2, "0");
+
+  return `${year}년 ${month}월 ${day}일 ${hours}:${minutes}`;
+};
